Implement matching when listing tasks
Filtering tasks by text/header can be helpful. And as useful as it
can be to use `--dir` to specify the base dir, sometimes you just
want to do it quicker, so you can also specify a dir or file as
an argument to restrict searching just to that.

Arguments can be mixed and matched freely. Multiple search terms
are additive not exclusive.